Fix bug where parent could resume without child being set
This is because the worker thread could call the enqueue block,
which then creates the child thread, but if the scheduler is letting
the child go, then the worker may not resume to assign the child.
Then, when the child wakes the parent, the scheduler may resume the parent
instead of the worker, thus the worker still has not assigned the child thread,
and the parent is resumed.

This is not a problem in prob, b/c it's due to a refactoring I just made.

The solution is to have the child thread set the variable rather than the worker thread.