Extract a ResumeThread object
It observes the child and wakes or raises in the parent
at the appropriate time.

Name the worker thread so that we can look at the thread name
for any relevant thread (well... the main thread doesn't have a name, I
guess :/)

Don't expose the worker to the outside world.